execution-time-estimator: separate library and aspect
Do not put the library and aspect into the plugin's directory, but
instead move them to the library directory.
