+++ src/libs/aspect/inifins/execution_time_estimator.cpp
@@ -18,7 +18,7 @@
  *  Read the full text in the LICENSE.GPL file in the doc directory.
  */
 
-#include "execution_time_estimator_aspect_inifin.h"
+#include "execution_time_estimator.h"
 
 namespace fawkes {
