+++ src/libs/aspect/inifins/execution_time_estimator.h
@@ -1,5 +1,5 @@
 /***************************************************************************
- *  execution_time_estimator_aspect_inifin.h - Aspect INiFin
+ *  execution_time_estimator.h - Aspect INiFin
  *
  *  Created: Thu 12 Dec 2019 19:00:07 CET 19:00
  *  Copyright  2019  Till Hofmann <hofmann@kbsg.rwth-aachen.de>
@@ -20,12 +20,13 @@
 
 #pragma once
 
-#include "execution_time_estimator_aspect.h"
+#include "../execution_time_estimator.h"
 
 #include <aspect/aspect.h>
 #include <aspect/inifins/inifin.h>
 
 namespace fawkes {
+
 class ExecutionTimeEstimatorsAspectIniFin : public virtual AspectIniFin
 {
 public:
