+++ src/libs/execution_time_estimator/Makefile
@@ -1,7 +1,7 @@
 #*****************************************************************************
 #                      Makefile Build System for Fawkes
 #                            -------------------
-#   Created on Thu 23 Apr 2020 16:53:02 CEST
+#   Created on Thu 23 Apr 2020 19:36:47 CEST
 #   Copyright (C) 2020 by Till Hofmann <hofmann@kbsg.rwth-aachen.de>
 #
 #*****************************************************************************
@@ -13,22 +13,20 @@
 #
 #*****************************************************************************
 
+
 BASEDIR = ../../..
-
 include $(BASEDIR)/etc/buildsys/config.mk
-include $(LIBSRCDIR)/utils/utils.mk
 
-LIBS_execution_time_estimator = fawkescore fawkes_execution_time_estimator
+LIBS_libfawkes_execution_time_estimator = fawkescore
+OBJS_libfawkes_execution_time_estimator = execution_time_estimator.o
 
-OBJS_execution_time_estimator = execution_time_estimator_plugin.o \
-                                execution_time_estimator_thread.o \
-                                estimators/config_estimator.o
+OBJS_all = $(OBJS_libfawkes_execution_time_estimator)
 
-OBJS_all = $(OBJS_execution_time_estimator)
-PLUGINS_all = $(PLUGINDIR)/execution-time-estimator.so
+LIBS_all = $(LIBDIR)/libfawkes_execution_time_estimator.so
 
 ifeq ($(HAVE_CPP17),1)
   CFLAGS        += $(CFLAGS_CPP17)
+  LIBS_build     = $(LIBS_all)
   PLUGINS_build  = $(PLUGINS_all)
 else
   WARN_TARGETS += warning_cpp17
@@ -39,7 +37,7 @@ all: $(WARN_TARGETS)
 .PHONY: $(WARN_TARGETS)
 
 warning_cpp17:
-	$(SILENT)echo -e "$(INDENT_PRINT)--> $(TRED)Omitting execution time estimator plugin$(TNORMAL) (C++17 not supported)"
+	$(SILENT)echo -e "$(INDENT_PRINT)--> $(TRED)Omitting exeuction time estimator$(TNORMAL) (C++17 not supported)"
 endif
 
 include $(BUILDSYSDIR)/base.mk
