+++ src/plugins/execution-time-estimator/execution_time_estimator_thread.h
@@ -20,11 +20,10 @@
 
 #pragma once
 
-#include "execution_time_estimator_aspect/execution_time_estimator_aspect.h"
-#include "execution_time_estimator_aspect/execution_time_estimator_aspect_inifin.h"
-
 #include <aspect/aspect_provider.h>
 #include <aspect/configurable.h>
+#include <aspect/execution_time_estimator.h>
+#include <aspect/inifins/execution_time_estimator.h>
 #include <core/threading/thread.h>
 
 class ExecutionTimeEstimatorsThread : public fawkes::Thread,
